+++ Pikartisan/resources/shaders/Renderer3D/DefaultStaticMeshShader.glsl
@@ -3,7 +3,7 @@
 layout(location = 0) in vec3 a_Position;
 layout(location = 1) in vec3 a_Normal;
 layout(location = 2) in vec2 a_TexCoord;
-layout(location = 3) in int a_EntityID;
+layout(location = 3) in highp int a_EntityID;
 
 layout(std140, binding = 0) uniform CameraData
 {
@@ -12,7 +12,7 @@ layout(std140, binding = 0) uniform CameraData
 	mat4 u_ProjectionMatrix;
 };
 
-out flat int v_EntityID;
+out flat highp int v_EntityID;
 
 void main() {
 	v_EntityID = a_EntityID;
@@ -23,9 +23,9 @@ void main() {
 #FRAGMENT_BEGIN()
 #version 460 core
 layout(location = 0) out vec4 o_FragmentColor;
-layout(location = 1) out int o_EntityID;
+layout(location = 1) out highp int o_EntityID;
 
-in flat int v_EntityID;
+in flat highp int v_EntityID;
 
 void main() {
 	o_FragmentColor = vec4(1.0f, 0.0f, 1.0f, 1.0f);
